[ADD] Confirmation pop delete new

+++ src/components/news/NewsDetail.js
@@ -1,6 +1,7 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link, useHistory, useParams } from "react-router-dom";
+import Swal from "sweetalert2";
 import { eventStartDeleted, newStartDetailLoading } from "../../actions/news";
 import { statsCategoryConsumed } from "../../actions/stats";
 import trashcan from "../../assets/trash-can.svg";
@@ -20,8 +21,20 @@ export default function NewsDetail() {
   }, [dispatch, _id]);
 
   const handleDeleteNew = () => {
-    dispatch(eventStartDeleted(_id));
-    history.replace('/');
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(eventStartDeleted(_id));
+        history.replace('/');
+      }
+    });
   };
 
   const { newSelected } = useSelector((state) => state.news);
